Moved tasks to inner class of test workflow
